+++ zia/provider_sweeper_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/firewallpolicies/networkservices"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/forwarding_control_policy/forwarding_rules"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/ips_control_policies/ips_policies"
+	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/ips_control_policies/ips_signature_rules"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/location/locationmanagement"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/rule_labels"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zia/services/sandbox/sandbox_rules"
@@ -96,6 +97,7 @@ func TestRunForcedSweeper(t *testing.T) {
 	}
 
 	sweepTestRuleLabels(testClient)
+	sweepTestIPSSignatureRules(testClient)
 	// sweepTestSourceIPGroup(testClient)
 	sweepTestDestinationIPGroup(testClient)
 	sweepTestNetworkServices(testClient)
@@ -163,6 +165,35 @@ func sweepTestRuleLabels(client *testClient) error {
 		}
 	}
 	// Log errors encountered during the deletion process
+	if len(errorList) > 0 {
+		for _, err := range errorList {
+			sweeperLogger.Error(err.Error())
+		}
+	}
+	return condenseError(errorList)
+}
+
+func sweepTestIPSSignatureRules(client *testClient) error {
+	var errorList []error
+
+	service := &zscaler.Service{
+		Client: client.sdkV3Client,
+	}
+
+	signatures, err := ips_signature_rules.GetAll(context.Background(), service)
+	if err != nil {
+		return err
+	}
+	sweeperLogger.Warn(fmt.Sprintf("Found %d resources to sweep", len(signatures)))
+	for _, s := range signatures {
+		if strings.HasPrefix(s.Name, testResourcePrefix) || strings.HasPrefix(s.Name, updateResourcePrefix) {
+			if _, err := ips_signature_rules.Delete(context.Background(), service, s.ID); err != nil {
+				errorList = append(errorList, err)
+				continue
+			}
+			logSweptResource(resourcetype.IPSSignatureRules, fmt.Sprintf("%d", s.ID), s.Name)
+		}
+	}
 	if len(errorList) > 0 {
 		for _, err := range errorList {
 			sweeperLogger.Error(err.Error())